requirements.txt
fastapi
uvicorn
opencv-python
pydantic-settings
numpy
pillow
torch
tensorflow
facenet-pytorch
insightface
onnxruntime
dlib
scikit-image
python-multipart

